Refactor chess piece classes: move PieceColor and Piece definitions to pieces.py; clean up models.py
